requirements.txt
PySide6>=6.10.1
Pillow>=10.4.0
watchdog==4.0.2
pypdf>=5.0.0